function double => float with a bad cast because Math func are double

+++ GGJ/Assets/Waves.Generator/base.cs
@@ -10,12 +10,12 @@ namespace Assets.Waves.Generator
         public Base() {
         }
         
-        public double Function(double input)
+        public float Function(float input)
         {
             return InternalFunction(input);
         }
 
-        virtual public double InternalFunction(double x) { return x; }
+        virtual public float InternalFunction(float x) { return x; }
     }
 
 }
